fix compilation warnings
Former-commit-id: 248c7e5cbc012ce915e23e60b18b528958915b24

+++ contrib/applications/platform/P300Stimulator/src/handlers/ovexP300ResultAreaHandler.cpp
@@ -173,7 +173,7 @@ void P300ResultAreaHandler::update(GObservable* observable, const void * pUserDa
 					m_sSpelledLetters.clear();
 			}
 
-			l_i32UndoSize = l_i32UndoSize<=m_ui32ResultCounter?l_i32UndoSize:m_ui32ResultCounter;
+			l_i32UndoSize = (uint32)l_i32UndoSize<=m_ui32ResultCounter?l_i32UndoSize:m_ui32ResultCounter;
 			for (uint32 i=1; i<=(uint32)l_i32UndoSize; i++)
 			{
 				m_pSymbolContainer->removeChild(m_ui32ResultCounter-i);
